refactor(cable): split cable related code into separate submodules
