+++ television/channels/cable.rs
@@ -1,33 +1,20 @@
-use rustc_hash::FxHashMap;
-use std::{
-    fmt::{self, Display, Formatter},
-    ops::Deref,
-};
-
 use std::collections::HashSet;
 use std::io::{BufRead, BufReader};
 use std::process::Stdio;
 
-use anyhow::Result;
-use regex::Regex;
+use preview::{parse_preview_kind, PreviewKind};
+use prototypes::{CableChannelPrototype, DEFAULT_DELIMITER};
 use rustc_hash::{FxBuildHasher, FxHashSet};
 use tracing::debug;
 
+use crate::channels::entry::{Entry, PreviewCommand, PreviewType};
 use crate::channels::OnAir;
 use crate::matcher::Matcher;
 use crate::matcher::{config::Config, injector::Injector};
 use crate::utils::command::shell_command;
-use crate::{
-    cable::ChannelPrototypes,
-    channels::entry::{Entry, PreviewCommand, PreviewType},
-};
 
-#[derive(Debug, Clone, PartialEq)]
-pub enum PreviewKind {
-    Command(PreviewCommand),
-    Builtin(PreviewType),
-    None,
-}
+pub mod preview;
+pub mod prototypes;
 
 #[allow(dead_code)]
 pub struct Channel {
@@ -66,17 +53,6 @@ impl From<CableChannelPrototype> for Channel {
                 None => None,
             },
         )
-    }
-}
-
-pub fn parse_preview_kind(command: &PreviewCommand) -> Result<PreviewKind> {
-    debug!("Parsing preview kind for command: {:?}", command);
-    let re = Regex::new(r"^\:(\w+)\:$").unwrap();
-    if let Some(captures) = re.captures(&command.command) {
-        let preview_type = PreviewType::try_from(&captures[1])?;
-        Ok(PreviewKind::Builtin(preview_type))
-    } else {
-        Ok(PreviewKind::Command(command.clone()))
     }
 }
 
@@ -232,96 +208,5 @@ impl OnAir for Channel {
 
     fn supports_preview(&self) -> bool {
         self.preview_kind != PreviewKind::None
-    }
-}
-
-#[derive(Clone, Debug, serde::Deserialize, PartialEq)]
-pub struct CableChannelPrototype {
-    pub name: String,
-    pub source_command: String,
-    #[serde(default)]
-    pub interactive: bool,
-    pub preview_command: Option<String>,
-    #[serde(default = "default_delimiter")]
-    pub preview_delimiter: Option<String>,
-}
-
-impl CableChannelPrototype {
-    pub fn new(
-        name: &str,
-        source_command: &str,
-        interactive: bool,
-        preview_command: Option<String>,
-        preview_delimiter: Option<String>,
-    ) -> Self {
-        Self {
-            name: name.to_string(),
-            source_command: source_command.to_string(),
-            interactive,
-            preview_command,
-            preview_delimiter,
-        }
-    }
-}
-
-const DEFAULT_PROTOTYPE_NAME: &str = "files";
-const DEFAULT_SOURCE_COMMAND: &str = "fd -t f";
-const DEFAULT_PREVIEW_COMMAND: &str = ":files:";
-
-impl Default for CableChannelPrototype {
-    fn default() -> Self {
-        Self {
-            name: DEFAULT_PROTOTYPE_NAME.to_string(),
-            source_command: DEFAULT_SOURCE_COMMAND.to_string(),
-            interactive: false,
-            preview_command: Some(DEFAULT_PREVIEW_COMMAND.to_string()),
-            preview_delimiter: Some(DEFAULT_DELIMITER.to_string()),
-        }
-    }
-}
-
-pub const DEFAULT_DELIMITER: &str = " ";
-
-#[allow(clippy::unnecessary_wraps)]
-fn default_delimiter() -> Option<String> {
-    Some(DEFAULT_DELIMITER.to_string())
-}
-
-impl Display for CableChannelPrototype {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        write!(f, "{}", self.name)
-    }
-}
-
-#[derive(Debug, serde::Deserialize)]
-pub struct CableChannels(pub FxHashMap<String, CableChannelPrototype>);
-
-impl Deref for CableChannels {
-    type Target = FxHashMap<String, CableChannelPrototype>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-#[cfg(unix)]
-const DEFAULT_CABLE_CHANNELS_FILE: &str =
-    include_str!("../../cable/unix-channels.toml");
-#[cfg(not(unix))]
-const DEFAULT_CABLE_CHANNELS_FILE: &str =
-    include_str!("../../cable/windows-channels.toml");
-
-impl Default for CableChannels {
-    /// Fallback to the default cable channels specification (the template file
-    /// included in the repo).
-    fn default() -> Self {
-        let pts =
-            toml::from_str::<ChannelPrototypes>(DEFAULT_CABLE_CHANNELS_FILE)
-                .expect("Unable to parse default cable channels");
-        let mut channels = FxHashMap::default();
-        for prototype in pts.prototypes {
-            channels.insert(prototype.name.clone(), prototype);
-        }
-        CableChannels(channels)
     }
 }
