+++ television/channels/remote_control.rs
@@ -1,6 +1,6 @@
 use std::collections::HashSet;
 
-use crate::channels::cable::CableChannels;
+use crate::channels::cable::prototypes::CableChannels;
 use crate::channels::entry::{Entry, PreviewType};
 use crate::channels::{OnAir, TelevisionChannel};
 use crate::matcher::{config::Config, Matcher};
